stream [nfc]: Split up StreamList into a copy for each caller
This gives us the first step toward fixing #3767.

A lot of the complexity of this cluster of components comes from the
fact that the two call sites of StreamList are really using it to do
different things.  Instead, we'll have them each do the things they
actually want to do.

At this first step we're duplicating code, so at first glance it
seems bad.  But, next, we'll make a number of simplifications that
become possible only because we've split these up.

In fact, as one measure of how much complexity comes from attempting
to share this code between these two use cases, when all is said and
done, the total diffstat for this branch is actually quite *negative*:
 src/streams/StreamItem.js           |   3 +-
 src/streams/StreamList.js           | 105 ----------------------------
 src/streams/SubscriptionsCard.js    |  49 +++++++++++--
 src/subscriptions/StreamListCard.js |  64 ++++++++++++-----
 4 files changed, 95 insertions(+), 126 deletions(-)

A split like this could be good even if in the end it meant more
total code, if in exchange the new code was simpler and easier to
modify than the old.  But in this case, there isn't even a trade-off.
